+++ src/activitypub/valueflows/constants.rs
@@ -3,6 +3,7 @@ pub const ACTION_DELIVER_SERVICE: &str = "deliverService";
 pub const ACTION_TRANSFER: &str = "transfer";
 
 pub const CLASS_CONTENT: &str = "https://www.wikidata.org/wiki/Q1260632";
+pub const CLASS_USER_GENERATED_CONTENT: &str = "https://www.wikidata.org/wiki/Q579716";
 
 // http://www.ontology-of-units-of-measure.org/resource/om-2/one
 pub const UNIT_ONE: &str = "one";