+++ src/activitypub/valueflows/parsers.rs
@@ -12,6 +12,7 @@ use super::constants::{
     ACTION_DELIVER_SERVICE,
     ACTION_TRANSFER,
     CLASS_CONTENT,
+    CLASS_USER_GENERATED_CONTENT,
     UNIT_ONE,
     UNIT_SECOND,
 };
@@ -77,8 +78,12 @@ pub fn parse_proposal(
     if proposal.publishes.action != ACTION_DELIVER_SERVICE {
         return Err(ValidationError("unexpected action"));
     };
-    if proposal.publishes.resource_conforms_to != CLASS_CONTENT {
-        return Err(ValidationError("unexpected resource type"));
+    // Use resource class to determine support for FEP-0837
+    // TODO: ignore non-interactive proposals (pre FEP-0837)
+    let fep_0837_enabled = match proposal.publishes.resource_conforms_to.as_str() {
+        CLASS_CONTENT => false,
+        CLASS_USER_GENERATED_CONTENT => true,
+        _ => return Err(ValidationError("unexpected resource type")),
     };
     let duration = proposal.publishes.resource_quantity
         .parse_duration()?;
@@ -104,6 +109,7 @@ pub fn parse_proposal(
         asset_type.chain_id,
         price,
         proposal.id,
+        fep_0837_enabled,
     );
     Ok(payment_option)
 }